requirements.txt
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
dotenv==0.9.9
exceptiongroup==1.3.1
fonttools==4.60.2
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.27
langchain-core==0.3.83
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langfuse==3.7.0
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
matplotlib==3.9.4
numpy==2.0.2
ollama==0.6.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pillow==11.3.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
SQLAlchemy==2.0.46
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
